src/main/java/com/juaracoding/Employee.java
package com.juaracoding;

public class Employee {
    // name, address, division, salary
    private String name;
    private String address;
    private boolean isActive;
    private double salary;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public boolean isActive() {
        return isActive;
    }

    public void setAcitve(boolean active) {
        isActive = active;
    }

    public double getSalary() {
        return salary;
    }

    public void setSalary(double salary) {
        if (salary > 0) {
          this.salary = salary;
        } else {
            throw new IllegalArgumentException("Salary harus lebih dari 0");
        }
    }
}
